chore(bidirectional-tool): Clean up longest/shortest diameters calculation
